Lib

Description

The lib of the project. Most of the work is done here.

Documentation

getEnding :: FilePath -> String

Extracts the file extension from a path.

getFileName :: FilePath -> String

Extracts the last part of the path, so the actual filename.

filterFileTyp :: String -> [FilePath] -> [FilePath]

Filters the files without the proper ending.

filterDirs :: [FilePath] -> IO [FilePath]

Only returns a list of the directories.

copyFileToDst :: FilePath -> FilePath -> IO ()

Copies the file to the specified location.

copyDir :: FilePath -> FilePath -> IO ()

Inspired by StackOverflow. Copies a whole Dir. Destination directory will be created. | Fails if the destination directory already exists.

filteredCopy :: FilePath -> FilePath -> String -> IO ()

Copies all the files with a specific ending from src to dst.

multiCopy :: FilePath -> FilePath -> Int -> Int -> IO ()

Copies a file multiple times. It will be named after the iteration.

massMultiCopy :: FilePath -> FilePath -> String -> Int -> IO ()

Makes multiple copies of the specified filetype. Absolutely unique feature.

multiCopyCaller :: ((FilePath, FilePath), (Int, Int)) -> IO ()

Helps to call multiCopy with the appropriate values.